# openapi/pulse.yaml
openapi: 3.1.0
info:
  title: Pulse API
  version: 0.1.0
  description: |
    Pulse V2 — Personal Life OS backend.
    **Spec-first**: add every endpoint here before implementing it.
servers:
  - url: http://localhost:4000
    description: Local dev
  - url: https://api.pulse.app
    description: Production (Render)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/teams:
    post:
      tags: [Teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                description: { type: string }
                avatar_url:  { type: string, format: uri }
      responses:
        '201': { description: Created }
    get:
      tags: [Teams]
      summary: List my teams
      responses:
        '200': { description: OK }
  /api/teams/{id}:
    get:
      tags: [Teams]
      summary: Get team detail + members
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK }
    patch:
      tags: [Teams]
      summary: Update team (owner only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:        { type: string }
                description: { type: string }
      responses:
        '200': { description: OK }
    delete:
      tags: [Teams]
      summary: Delete team (owner only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Deleted }
  /api/teams/{id}/invite:
    post:
      tags: [Teams]
      summary: Invite a user to the team
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
                role:    { type: string, enum: [admin, member] }
      responses:
        '201': { description: Invited }
  /api/teams/{id}/members/{userId}:
    delete:
      tags: [Teams]
      summary: Remove a member
      parameters:
        - { name: id,     in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Removed }
    patch:
      tags: [Teams]
      summary: Update member role
      parameters:
        - { name: id,     in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [admin, member] }
      responses:
        '200': { description: Updated }
  /api/teams/{id}/leave:
    post:
      tags: [Teams]
      summary: Leave a team
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Left }

  # ── Organizations ───────────────────────────────────────────────────────────
...
  /api/shares/{entityType}/{entityId}:
    post:
      tags: [Shares]
      summary: Share an entity with a user or team
      parameters:
        - { name: entityType, in: path, required: true, schema: { type: string } }
        - { name: entityId,   in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permission]
              properties:
                permission:   { type: string, enum: [view, comment, edit] }
                with_user_id: { type: string, format: uuid }
                with_team_id: { type: string, format: uuid }
      responses:
        '201': { description: Created }
    get:
      tags: [Shares]
      summary: List shares for an entity
      parameters:
        - { name: entityType, in: path, required: true, schema: { type: string } }
        - { name: entityId,   in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK }
  /api/shares/{entityType}/{entityId}/{shareId}:
    patch:
      tags: [Shares]
      summary: Update share permission
      parameters:
        - { name: entityType, in: path, required: true, schema: { type: string } }
        - { name: entityId,   in: path, required: true, schema: { type: string, format: uuid } }
        - { name: shareId,    in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permission]
              properties:
                permission: { type: string, enum: [view, comment, edit] }
      responses:
        '200': { description: OK }
    delete:
      tags: [Shares]
      summary: Revoke a share
      parameters:
        - { name: entityType, in: path, required: true, schema: { type: string } }
        - { name: entityId,   in: path, required: true, schema: { type: string, format: uuid } }
        - { name: shareId,    in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Revoked }

  # ── Notifications ────────────────────────────────────────────────────────────
...